+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## v3.15.0
 - Deprecate DataProviderNameFixer - use "php_unit_data_provider_name"
-- Deprecate PhpdocParamOrderFixer
+- Deprecate PhpdocParamOrderFixer - use "phpdoc_param_order"
 - Update minimum PHP CS Fixer version to 3.19.0
 
 ## v3.14.0